Revert "[#40]direct forwarding을 위한 RewritePath Filter routes 속성 적용 "
