--- client/http-client.env.json
@@ -1,7 +1,0 @@
-{
-  "local": {
-    "user-service": "http://localhost:8000/user-service",
-    "product-service": "http://localhost:8000/product-service",
-    "order-service": "http://localhost:8000/order-service"
-  }
-}
