--- client/product-service.http
@@ -1,11 +1,0 @@
-# 특정 상품 조회
-GET {{product-service}}/1
-Content-Type: application/json
-
-###
-
-# 상품 목록 조회
-GET {{product-service}}/product?page=0&size=10&sort=createdAt&direction=DESC
-Content-Type: application/json
-
-###
