--- client/user-service.http
@@ -1,26 +1,0 @@
-# 회원 가입
-POST {{user-service}}/user
-Content-Type: application/json
-
-{
-  "email": "test@gmail.com",
-  "password": "password",
-  "name": "choi-ys"
-}
-> {%
-client.global.set("signupUserId", response.body.userId)
-%}
-
-###
-
-# 특정 회원 조회
-GET {{user-service}}/user/{{signupUserId}}
-Content-Type: application/json
-
-###
-
-# 회원 목록 조회
-GET {{user-service}}/user?page=0&size=10&sort=createdAt&direction=DESC
-Content-Type: application/json
-
-###
