+++ product-service/src/main/kotlin/io/ecommerce/productgservice/ProductServiceApplication.kt
@@ -1,0 +1,13 @@
+package io.ecommerce.productgservice
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient
+
+@SpringBootApplication
+@EnableDiscoveryClient
+class ProductServiceApplication
+
+fun main(args: Array<String>) {
+    runApplication<ProductServiceApplication>(*args)
+}
