+++ product-service/src/main/kotlin/io/ecommerce/productgservice/core/domain/common/Auditor.kt
@@ -1,0 +1,26 @@
+package io.ecommerce.productgservice.core.domain.common
+
+import org.hibernate.annotations.CreationTimestamp
+import org.hibernate.annotations.UpdateTimestamp
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+import javax.persistence.*
+
+/**
+ * @author : choi-ys
+ * @date : 2022/01/08 7:08 오후
+ */
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener::class)
+abstract class Auditor(
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long = 0L,
+
+    @CreationTimestamp
+    @Column(name = "created_at")
+    var createdAt: LocalDateTime? = null,
+
+    @UpdateTimestamp
+    @Column(name = "updated_at")
+    var updatedAt: LocalDateTime? = null,
+)
