--- product-service/src/main/kotlin/io/ecommerce/productservice/ProductServiceApplication.kt
@@ -1,13 +1,0 @@
-package io.ecommerce.productservice
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient
-
-@SpringBootApplication
-@EnableDiscoveryClient
-class ProductServiceApplication
-
-fun main(args: Array<String>) {
-    runApplication<ProductServiceApplication>(*args)
-}
