+++ product-service/src/test/kotlin/io/ecommerce/productgservice/ProductServiceApplicationTests.kt
@@ -1,0 +1,13 @@
+package io.ecommerce.productgservice
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ProductServiceApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
