--- product-service/src/test/kotlin/io/ecommerce/productservice/ProductServiceApplicationTests.kt
@@ -1,13 +1,0 @@
-package io.ecommerce.productservice
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ProductServiceApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
