--- product-service/src/test/kotlin/io/ecommerce/productservice/config/test/SpringBootTestSupporter.kt
@@ -1,38 +1,0 @@
-package io.ecommerce.productservice.config.test
-
-import org.junit.jupiter.api.TestInstance
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Import
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.TestConstructor
-import org.springframework.transaction.annotation.Transactional
-import org.springframework.web.filter.CharacterEncodingFilter
-import java.lang.annotation.*
-import java.lang.annotation.Retention
-import java.lang.annotation.Target
-
-/**
- * @author : choi-ys
- * @date : 2022/01/09 6:17 오후
- */
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@AutoConfigureMockMvc
-@Transactional
-@ActiveProfiles("test")
-@Import(SpringBootTestSupporter.Config::class)
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Inherited
-annotation class SpringBootTestSupporter() {
-    class Config {
-        @Bean
-        fun characterEncodingFilter(): CharacterEncodingFilter {
-            return CharacterEncodingFilter("UTF-8", true)
-        }
-    }
-}
